+++ src/xcore/bookmarks.cpp
@@ -1,3 +1,6 @@
+#include <QFileInfo>
+#include <QDir>
+
 #include "xcore.h"
 
 void addBookmark(std::string nm, std::string fp) {
@@ -5,6 +8,26 @@ void addBookmark(std::string nm, std::string fp) {
 	nbm.name = nm;
 	nbm.path = fp;
 	conf.bookmarkList.push_back(nbm);
+}
+
+// one file whichever way its path was written; no disk access, so a favorite
+// on a drive that is gone cannot hold the menu up
+static QString bkm_key(const QString& path) {
+	return QDir::cleanPath(QFileInfo(path).absoluteFilePath());
+}
+
+int findBookmark(const QString& path) {
+#ifdef _WIN32
+	Qt::CaseSensitivity cs = Qt::CaseInsensitive;
+#else
+	Qt::CaseSensitivity cs = Qt::CaseSensitive;
+#endif
+	QString key = bkm_key(path);
+	for (int i = 0; i < conf.bookmarkList.size(); i++) {
+		if (bkm_key(QString::fromLocal8Bit(conf.bookmarkList[i].path.c_str())).compare(key, cs) == 0)
+			return i;
+	}
+	return -1;
 }
 
 void swapBookmarks(int p1, int p2) {